Fix ANMultiAdRequest corner case where fetcher is not recreated.
REVIEWERS
  Primary:    Abhishek, Akash, Kowshick, Punnaghai
  Secondary:  Abhas, Wei

DESCRIPTION

This update handles the case where an AdUnit derived from ANAdView executes loadAd before being included in an instance of ANMultiAdRequest.
Or vice versa, where the AdUnit is removed from the MAR instance after being loaded once by the MAR instance.

The problem is that the fetcher is defined at the time of the load, and the definition is specific to whether the internal property
marManager is defined.

The solution is to stop and free the fetcher whenever marManager is assigned.